app/requirements.txt
Flask==2.3.3
Cerberus==1.3.4
pytest-cov==5.0.0
python-dotenv==1.0.1